backend/pyproject.toml
[tool.poetry]
name = "koop"
version = "0.1.0"
description = ""
authors = ["Natalia <natalia@example.com>", "Jerzy <jerzy@example.com>"]
license = "GPL-3.0 License"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
psycopg2-binary = "^2.8"
django-cors-headers = "^3.7"
djangorestframework = "^3.12"

[tool.poetry.dev-dependencies]
black = {version = "*", allow-prereleases = true}
django-debug-toolbar = "^3.2"
ipython = "^7.22"
pre-commit = "^2.12"
mypy = "^0.812"
djangorestframework-stubs = "^1.4.0"
factory-boy = "^3.2.0"

[tool.black]
target-version = ['py38']
exclude = 'migrations/|koop/settings'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
